api/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile --output-file=api/requirements.txt api/requirements.in
#
amqp==5.3.1
    # via kombu
annotated-doc==0.0.4
    # via fastapi
annotated-types==0.7.0
    # via pydantic
anyio==4.12.0
    # via
    #   httpx
    #   starlette
    #   watchfiles
asgiref==3.11.0
    # via opentelemetry-instrumentation-asgi
billiard==4.2.4
    # via celery
cachetools==6.2.4
    # via google-auth
celery==5.6.0
    # via -r api/requirements.in
certifi==2025.11.12
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==2.0.0
    # via cryptography
charset-normalizer==3.4.4
    # via requests
click==8.3.1
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
    #   uvicorn
click-didyoumean==0.3.1
    # via celery
click-plugins==1.1.1.2
    # via celery
click-repl==0.3.0
    # via celery
colorama==0.4.6
    # via
    #   click
    #   pytest
    #   uvicorn
cryptography==46.0.3
    # via pyjwt
exceptiongroup==1.3.1
    # via celery
fastapi==0.128.0
    # via -r api/requirements.in
google-auth==2.45.0
    # via -r api/requirements.in
googleapis-common-protos==1.72.0
    # via
    #   opentelemetry-exporter-otlp-proto-grpc
    #   opentelemetry-exporter-otlp-proto-http
grpcio==1.76.0
    # via opentelemetry-exporter-otlp-proto-grpc
gunicorn==23.0.0
    # via -r api/requirements.in
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.9
    # via httpx
httptools==0.7.1
    # via uvicorn
httpx==0.28.1
    # via -r api/requirements.in
idna==3.11
    # via
    #   anyio
    #   httpx
    #   requests
importlib-metadata==8.7.1
    # via opentelemetry-api
iniconfig==2.3.0
    # via pytest
kombu==5.6.1
    # via celery
opentelemetry-api==1.39.1
    # via
    #   -r api/requirements.in
    #   opentelemetry-exporter-otlp-proto-grpc
    #   opentelemetry-exporter-otlp-proto-http
    #   opentelemetry-instrumentation
    #   opentelemetry-instrumentation-asgi
    #   opentelemetry-instrumentation-dbapi
    #   opentelemetry-instrumentation-fastapi
    #   opentelemetry-instrumentation-psycopg2
    #   opentelemetry-sdk
    #   opentelemetry-semantic-conventions
opentelemetry-exporter-otlp==1.39.1
    # via -r api/requirements.in
opentelemetry-exporter-otlp-proto-common==1.39.1
    # via
    #   opentelemetry-exporter-otlp-proto-grpc
    #   opentelemetry-exporter-otlp-proto-http
opentelemetry-exporter-otlp-proto-grpc==1.39.1
    # via opentelemetry-exporter-otlp
opentelemetry-exporter-otlp-proto-http==1.39.1
    # via opentelemetry-exporter-otlp
opentelemetry-instrumentation==0.60b1
    # via
    #   opentelemetry-instrumentation-asgi
    #   opentelemetry-instrumentation-dbapi
    #   opentelemetry-instrumentation-fastapi
    #   opentelemetry-instrumentation-psycopg2
opentelemetry-instrumentation-asgi==0.60b1
    # via opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-dbapi==0.60b1
    # via opentelemetry-instrumentation-psycopg2
opentelemetry-instrumentation-fastapi==0.60b1
    # via -r api/requirements.in
opentelemetry-instrumentation-psycopg2==0.60b1
    # via -r api/requirements.in
opentelemetry-proto==1.39.1
    # via
    #   opentelemetry-exporter-otlp-proto-common
    #   opentelemetry-exporter-otlp-proto-grpc
    #   opentelemetry-exporter-otlp-proto-http
opentelemetry-sdk==1.39.1
    # via
    #   -r api/requirements.in
    #   opentelemetry-exporter-otlp-proto-grpc
    #   opentelemetry-exporter-otlp-proto-http
opentelemetry-semantic-conventions==0.60b1
    # via
    #   opentelemetry-instrumentation
    #   opentelemetry-instrumentation-asgi
    #   opentelemetry-instrumentation-dbapi
    #   opentelemetry-instrumentation-fastapi
    #   opentelemetry-sdk
opentelemetry-util-http==0.60b1
    # via
    #   opentelemetry-instrumentation-asgi
    #   opentelemetry-instrumentation-fastapi
packaging==25.0
    # via
    #   gunicorn
    #   kombu
    #   opentelemetry-instrumentation
    #   pytest
passlib==1.7.4
    # via -r api/requirements.in
pluggy==1.6.0
    # via pytest
prompt-toolkit==3.0.52
    # via click-repl
protobuf==6.33.2
    # via
    #   googleapis-common-protos
    #   opentelemetry-proto
psycopg2-binary==2.9.11
    # via -r api/requirements.in
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.2
    # via google-auth
pycparser==2.23
    # via cffi
pydantic==2.12.5
    # via
    #   -r api/requirements.in
    #   fastapi
pydantic-settings==2.6.1
    # via -r api/requirements.in
pydantic-core==2.41.5
    # via pydantic
pygments==2.19.2
    # via pytest
pyjwt[crypto]==2.10.1
    # via -r api/requirements.in
pytest==9.0.2
    # via
    #   -r api/requirements.in
    #   pytest-asyncio
pytest-cov==5.0.0
    # via -r api/requirements.in
pytest-asyncio==1.3.0
    # via -r api/requirements.in
python-dateutil==2.9.0.post0
    # via celery
python-dotenv==1.2.1
    # via uvicorn
pyyaml==6.0.3
    # via uvicorn
types-pyyaml==6.0.12.20241230
    # via mypy type checking for tests using yaml
redis==7.1.0
    # via -r api/requirements.in
requests==2.32.5
    # via opentelemetry-exporter-otlp-proto-http
rsa==4.9.1
    # via google-auth
six==1.17.0
    # via python-dateutil
starlette==0.50.0
    # via fastapi
typing-extensions==4.15.0
    # via
    #   anyio
    #   exceptiongroup
    #   fastapi
    #   grpcio
    #   opentelemetry-api
    #   opentelemetry-exporter-otlp-proto-grpc
    #   opentelemetry-exporter-otlp-proto-http
    #   opentelemetry-sdk
    #   opentelemetry-semantic-conventions
    #   pydantic
    #   pydantic-core
    #   pytest-asyncio
    #   starlette
    #   typing-inspection
typing-inspection==0.4.2
    # via pydantic
tzdata==2025.3
    # via
    #   kombu
    #   tzlocal
tzlocal==5.3.1
    # via celery
urllib3==2.6.2
    # via requests
uvicorn[standard]==0.40.0
    # via -r api/requirements.in
vine==5.1.0
    # via
    #   amqp
    #   celery
    #   kombu
watchfiles==1.1.1
    # via uvicorn
wcwidth==0.2.14
    # via prompt-toolkit
websockets==15.0.1
    # via uvicorn
wrapt==1.17.3
    # via
    #   opentelemetry-instrumentation
    #   opentelemetry-instrumentation-dbapi
zipp==3.23.0
    # via importlib-metadata
ruff==0.6.9
    # dev linting
mypy==1.11.2
    # dev type checking
